US125660/move right panel skeleton to its own file.
