+++ test/perceptual/consistent-evaluation-grade-result.visual-diff.html
@@ -41,9 +41,5 @@
 		<div id="grade-letter">
 			<d2l-consistent-evaluation-grade-result/>
 		</div>
-
-		<div id="skeleton">
-			<d2l-consistent-evaluation-grade-result skeleton />
-		</div>
 	</body>
 </html>
