+++ test/perceptual/consistent-evaluation-outcomes.visual-diff.html
@@ -39,14 +39,5 @@
             ></d2l-consistent-evaluation-outcomes>
 		</div>
 
-		<div id="skeleton">
-			<d2l-consistent-evaluation-outcomes
-                    header="outcomes"
-					href="data/outcomes/suggested-level.json"
-					token="token"
-					skeleton
-            ></d2l-consistent-evaluation-outcomes>
-        </div>
-
 	</body>
 </html>
