+++ test/perceptual/consistent-evaluation-outcomes.visual-diff.js
@@ -28,9 +28,4 @@ describe('d2l-consistent-evaluation-outcomes', () => {
 		const rect = await visualDiff.getRect(page, '#outcome-with-description');
 		await visualDiff.screenshotAndCompare(page, this.test.fullTitle(), { clip: rect });
 	});
-
-	it('renders skeleton outcomes', async function() {
-		const rect = await visualDiff.getRect(page, '#skeleton');
-		await visualDiff.screenshotAndCompare(page, this.test.fullTitle(), { clip: rect });
-	});
 });
