+++ test/perceptual/consistent-evaluation-right-panel.visual-diff.html
@@ -49,6 +49,13 @@
 			></consistent-evaluation-right-panel>
 		</div>
 
+		<div id="renders-skeleton">
+			<consistent-evaluation-right-panel
+				skeleton
+				token="token"
+			></consistent-evaluation-right-panel>
+		</div>
+
 		<div id="rubric-read-only">
 			<consistent-evaluation-right-panel
 				allow-evaluation-write
