+++ test/perceptual/consistent-evaluation-rubric.visual-diff.html
@@ -58,13 +58,5 @@
 			></d2l-consistent-evaluation-rubric>
 		</div>
 
-		<div id="skeleton">
-			<d2l-consistent-evaluation-rubric
-				header="Rubrics"
-				token="token"
-				skeleton
-			></d2l-consistent-evaluation-rubric>
-		</div>
-
 	</body>
 </html>
